verify_email.py
The "verify_email.py" code is a Python script that allows you to verify the validity of a list of email addresses. The script extracts the domain name from each email address and queries the DNS for the domain's MX records. It then connects to each mail server and attempts to verify the email address by sending a series of commands. If the email address is verified by any of the mail servers, it is considered valid and written to an output file.
This script can be useful for validating email addresses in bulk for email marketing campaigns, newsletter subscriptions, or other email-based services. It can also be a helpful tool for cleaning up a mailing list and removing invalid email addresses.

Feel free to use and modify this script for your own purposes, and please note that it is for educational purposes only. If you use this script for any commercial purposes, be sure to comply with all applicable laws and regulations.